# app-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Notatek",
        "description": "Dokumentacja API dla systemu notatek",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Pobieranie listy notatek użytkownika",
                "operationId": "ea5a92fa2c7665842a92c619f2301f25",
                "responses": {
                    "200": {
                        "description": "Lista notatek"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Tworzenie nowej notatki",
                "operationId": "bb27bc018733c2dc72e7efd0713b0dac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nowa notatka"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Treść notatki"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notatka została utworzona"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notes/{id}": {
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Edycja notatki",
                "operationId": "2566aae29ac3263034f7f28bcad65cb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID notatki do edycji",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Zaktualizowana notatka"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Nowa treść notatki"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notatka została zaktualizowana"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Usuwanie notatki",
                "operationId": "a7afeb471a57034411e5831af62ce11e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID notatki do usunięcia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notatka została usunięta"
                    },
                    "404": {
                        "description": "Nie znaleziono notatki"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notes/shared": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Pobieranie współdzielonych notatek",
                "operationId": "9f242983783f1672314b6b9c7c705a21",
                "responses": {
                    "200": {
                        "description": "Lista współdzielonych notatek"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Wpisz token w formacie (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}